Delete useCreateGroupUserList.ts and add selectGroupUserList to function.ts
