+++ src/containers/home/transactionInputForm/InputFormContainer.tsx
@@ -21,7 +21,7 @@ import { GroupTransactionsReq } from '../../../reducks/groupTransactions/types';
 import { isValidAmountFormat } from '../../../lib/validation';
 import { year, customMonth } from '../../../lib/constant';
 import InputForm from '../../../components/home/transactionInputForm/InputForm';
-import { useCreateGroupUserList } from '../../../hooks/groupUsers/useCreateGroupUserList';
+import { selectGroupUserList } from '../../../lib/function';
 
 const initialState = {
   initialAmount: '',
@@ -62,9 +62,8 @@ const InputFormContainer = () => {
   const [editCustomCategoryName, setEditCustomCategoryName] = useState('');
   const [bigEditCategoryIndex, setBigEditCategoryIndex] = useState<number | null>(null);
   const [associatedIndex, setAssociatedIndex] = useState<number | null>(null);
-  const { createUsersList } = useCreateGroupUserList();
 
-  const selectUsersItemList = createUsersList(Number(group_id));
+  const selectUsersItemList = selectGroupUserList(approvedGroups, Number(group_id));
 
   const addTransactionDate = {
     addTransactionYear: 0,
